backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mangum==0.17.0
motor==3.4.0
pymongo[srv]==4.7.2
dnspython==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
httpx==0.27.0
python-multipart==0.0.9
